+++ azext_iot/commands.py
@@ -26,8 +26,6 @@ def load_command_table(self, _):
         cmd_group.command('monitor-events', 'iot_hub_monitor_events')
         cmd_group.command('monitor-feedback', 'iot_hub_monitor_feedback')
 
-        cmd_group.command('show-connection-string', 'iot_get_hub_connection_string')
-
     with self.command_group('iot hub device-identity', command_type=iothub_ops) as cmd_group:
         cmd_group.command('create', 'iot_device_create')
         cmd_group.command('show', 'iot_device_show')